+++ cli/server.go
@@ -37,6 +37,7 @@ func newServerCommand(cli *CLI) *cobra.Command {
 		newServerSSHCommand(cli),
 		newServerAddLabelCommand(cli),
 		newServerRemoveLabelCommand(cli),
+		newServerSetRdnsCommand(cli),
 	)
 	return cmd
 }