+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Changes
 
+## master
+* Add `hcloud ssh-key update` command
+
 ## v1.7.0
 
 * Add type filter flag `-t` / `--type` to `image list` command