+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changes
 
 ## master
+
 * Add `hcloud ssh-key update` command
 
 ## v1.7.0