+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## master
 
 * Add `hcloud ssh-key update` command
+* Add `hcloud server set-rdns` command
 
 ## v1.7.0