+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * Add `hcloud ssh-key update` command
 * Add `-u` & `-p` / `--port` to `server ssh` command
+
 ## v1.7.0
 
 * Add type filter flag `-t` / `--type` to `image list` command